requirements.txt
python-telegram-bot=12.3.0
bs4=4.8.4
requests=2.22.0

